Move sighash types to submodule
Move the two sighash types (`TapSighashType` and `EcdsaSighashType`) to
a `sighash_type` submodule.

While we are at it lay the code out with impl blocks below the type
definitions as is customary.

Code move, re-order, and fix import/export paths only. No other changes.
